Rename Ascend module directories to NPU
