+++ afd_plugin/compat/npu/feature_validation.py
