+++ afd_plugin/compat/npu/forward_context.py
