+++ afd_plugin/compat/npu/ops.py
