+++ afd_plugin/compat/npu/profiler.py
