+++ afd_plugin/compat/npu/runtime_config.py
