+++ afd_plugin/v1/worker/npu/attention_worker.py
@@ -9,13 +9,13 @@ from typing import Any
 from vllm.v1.worker.workspace import init_workspace_manager
 from vllm_ascend.worker.worker import NPUWorker
 
-from afd_plugin.compat.ascend import (
+from afd_plugin.compat.npu import (
     apply_afd_ascend_patches_if_needed,
     fail_if_unsupported_npu_afd_features,
     fix_all2all_backend_for_afd,
     npu_afd_num_ubatches,
 )
-from afd_plugin.v1.worker.ascend.attention_model_runner import (
+from afd_plugin.v1.worker.npu.attention_model_runner import (
     AFDNPUAttentionModelRunner,
 )
 from afd_plugin.validation import (
