+++ recipe/npu/CAMAsyncAFDConnector/deepseek_v3_2/README.md
@@ -235,7 +235,7 @@ ASCEND_RT_VISIBLE_DEVICES=0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15 \
 vllm serve /path/to/DeepSeek-V3.2 \
   --host 0.0.0.0 \
   --port 8000 \
-  --worker-cls afd_plugin.v1.worker.ascend.AFDNPUAttentionWorker \
+  --worker-cls afd_plugin.v1.worker.npu.AFDNPUAttentionWorker \
   --max-num-seqs 32 \
   --enforce-eager \
   --served-model-name deepseek_v3_2 \
@@ -301,7 +301,7 @@ ASCEND_RT_VISIBLE_DEVICES=0,1,2,3,4,5,6,7 \
 vllm serve /path/to/DeepSeek-V3.2 \
   --host 0.0.0.0 \
   --port 8000 \
-  --worker-cls afd_plugin.v1.worker.ascend.AFDNPUAttentionWorker \
+  --worker-cls afd_plugin.v1.worker.npu.AFDNPUAttentionWorker \
   --max-num-seqs 32 \
   --served-model-name deepseek_v3_2 \
   --enforce-eager \
@@ -367,7 +367,7 @@ export AFD_FORCE_BALANCED_TOPK_IDS=1
 ASCEND_RT_VISIBLE_DEVICES=8,9,10,11,12,13,14,15 \
 vllm serve /path/to/DeepSeek-V3.2 \
   --port 8001 \
-  --worker-cls afd_plugin.v1.worker.ascend.AFDNPUFFNWorker \
+  --worker-cls afd_plugin.v1.worker.npu.AFDNPUFFNWorker \
   --max-num-seqs 2 \
   --enforce-eager \
   --quantization ascend \
