+++ recipe/npu/CAMP2pAFDConnector/deepseek_v3_2/afd_attention.sh
@@ -105,7 +105,7 @@ KV_TRANSFER_CONFIG='{
 exec env VLLM_USE_V1=1 vllm serve "$MODEL_PATH" \
   --host 0.0.0.0 \
   --port "$SERVER_PORT" \
-  --worker-cls afd_plugin.v1.worker.ascend.AFDNPUAttentionWorker \
+  --worker-cls afd_plugin.v1.worker.npu.AFDNPUAttentionWorker \
   --data-parallel-size "$ATTENTION_RANKS" \
   --data-parallel-size-local 16 \
   --data-parallel-start-rank "$DP_START_RANK" \
